+++ elfpy/agents/__init__.py
@@ -1,3 +1,1 @@
 """Agent specifications and helper functions"""
-
-from .agent import Agent
