+++ elfpy/bots/bot_info.py
@@ -1,7 +1,8 @@
 """Information for creating a bot"""
 from __future__ import annotations
 
-from dataclasses import dataclass
+from collections import namedtuple
+from dataclasses import dataclass, field
 
 from elfpy.agents.policies.base import BasePolicy
 
@@ -33,6 +34,9 @@ class BotInfo:  # TODO: Rename to `BotConfig` when we remove evm_bots
     name: str = "BoringBotty"
     budget: Budget = Budget()
     number_of_bots: int = 1
+    init_kwargs: dict = field(default_factory=dict)
+    # TODO: Remove below once we remove evm_bots
     trade_chance: float = 0.8
-    risk_threshold: float = 0.8  # TODO: Remove this once we remove evm_bots
-    init_kwargs: dict = {}
+    risk_threshold: float = 0.8
+    Risk = namedtuple("Risk", ["mean", "std", "min", "max"])
+    risk: Risk = Risk(mean=0.02, std=0.01, min=0.0, max=0.06)
