+++ elfpy/eth/__init__.py
@@ -3,5 +3,5 @@
 from . import abi, accounts
 from .numeric_utils import convert_scaled_value
 from .rpc_interface import get_account_balance, set_anvil_account_balance
-from .transactions import fetch_transactions_for_block, smart_contract_read, smart_contract_transact
+from .transactions import smart_contract_read, smart_contract_transact
 from .web3_setup import initialize_web3_with_http_provider
